backend/requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
black==23.9.1
cfgv==3.4.0
click==8.1.7
distlib==0.3.7
Django==4.2.5
django-graphql-jwt==0.4.0
filelock==3.12.4
flake8==6.1.0
flake8-quotes==3.3.2
graphene==3.3
graphene-django==3.1.5
graphql-core==3.2.3
graphql-relay==3.2.0
identify==2.5.29
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pre-commit==3.4.0
promise==2.3
psycopg2-binary==2.9.9
pycodestyle==2.11.0
pyflakes==3.1.0
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
virtualenv==20.24.5
